PEP8 and refactoring.
