--- flock_controller/mechanics/listner.py
@@ -1,3 +1,0 @@
-"""Main control loop for central controller."""
-
-from flock_controller.main import
