+++ flock_controller/mechanics/location.py
@@ -1,8 +1,16 @@
 """Operations to update Area."""
 from flock_controller.mechanics.main import CENTRAL_SERVER
 from flock_controller.mechanics.main import RES_CS
-from flock_controller.mechanics.main import gen_Location
 from hydra import SCHEMA, Resource
+
+
+def gen_Location(coordinate_str):
+    """Generate a Location object."""
+    Area = {
+        "@type": "Location",
+        "Location": coordinate_str
+    }
+    return Area
 
 
 def update_location(location):
@@ -17,7 +25,7 @@ def update_location(location):
 
 
 if __name__ == "__main__":
-    ## Initialise central server location at 0,0
+    # Initialise central server location at 0,0
     location = gen_Location("0.856901647439813,14.08447265625")
     print(location)
