+++ flock_controller/mechanics/messages.py
@@ -2,8 +2,16 @@
 from hydra import Resource, SCHEMA
 from mechanics.main import CENTRAL_SERVER
 from mechanics.main import RES_CS
-from mechanics.main import gen_Message
 import json
+
+
+def gen_Message(message):
+    """Create a new Message."""
+    message = {
+        "@type": "Message",
+        "MessageString": message,
+    }
+    return message
 
 
 def get_message_collection():
