+++ flock_controller/mechanics/simulation.py
@@ -1,0 +1,3 @@
+"""Main control loop for central controller."""
+
+from flock_controller.main import get_drone, update_drone
